Update index.js and refactor all tests
